+++ cmd/init.go
@@ -27,12 +27,6 @@ var initCmd = &cobra.Command{
 			outputDirFlag += "/"
 		}
 
-		// Validate directories early
-		err = ValidateDirectories(inputDirFlag, outputDirFlag)
-		if err != nil {
-			log.Fatalln(err)
-		}
-
 		if len(valuesFileFlags) > 0 {
 			// Parse values from files first (merge multiple files)
 			values, err = parseValuesFromFiles(valuesFileFlags)